# backend/api_docs.yaml
...
openapi: 3.1.0
info:
  title: Backend service
  version: 0.1.0
paths:
  /v1/auth/authenticate:
    post:
      summary: Authenticate
      operationId: authenticate_v1_auth_authenticate_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Body_authenticate_v1_auth_authenticate_post"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/V1AuthAuthenticatePostResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
    Body_authenticate_v1_auth_authenticate_post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: ^password$
            - type: "null"
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ""
        client_id:
          anyOf:
            - type: string
            - type: "null"
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: "null"
          format: password
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_authenticate_v1_auth_authenticate_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    V1AuthAuthenticatePostResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
        - access_token
        - token_type
      title: V1AuthAuthenticatePostResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /v1/auth/authenticate